requirements.txt
# Telegram bot
aiogram==3.15.0

# SVG rendering
cairosvg==2.7.1

# Image processing
Pillow==11.1.0

# Templating
Jinja2==3.1.5

# Config
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.7.0
